Add affinity-assistant image
When attempting to use a pipeline with PVC workspace, tekton creates a
StatefulSet called affinity-assistant. Since the image wasn't defined,
it causes it to not be able to create a deployment. The taskruns that
are spun up are stuck in pending state as it couldn't be scheduled.
After a while the tasks will timeout and the build will fail.