Start to sketch our a `mink resolve` and `mink apply` commands.
These commands are akin to `ko resolve` and `ko apply`, but in place of directly `go build` and containerizing, it delegates to one of our supported build mechanisms.

In the following example, a single build "kontext" is produced for the repository (guided by `--directory`), and it is passed to 4 different builds:

```yaml
apiVersion: serving.knative.dev/v1
kind: Service
metadata:
  name: helloworld-go
  namespace: default
spec:
  template:
    spec:
      containers:
      - image: ko://helloworld-go
        env:
        - name: TARGET
          value: "Go Sample v1"
        - name: PYTHON
          value: buildpacks:///helloworld-python
        - name: NODE
          value: dockerfile:///helloworld-nodejs
        - name: SHELL
          value: dockerfile:///helloworld-shell
```

The main `image:` build is a `ko publish` invocation, which builds `helloworld-go` via `ko publish` (on-cluster).

The `PYTHON` build is a `mink buildpack` invocation, which runs over the source bundle with overrides drawn from
`./helloworld-python/overrides.toml` in the bundle.

The other two builds are `mink build` invocations, which run over the source with Dockerfiles drawn their
respective sub-directories.

This is still fairly naive with room for optimization, but is roughly functional.  Expect breaking changes as the contract here evolves.