fix: add an optional output directory for resolve
so that the yaml files are written to a directory and output the build logs

this lets you modify the yaml files in place on disk or output to a separate folder so that the yaml can then be processed on the file system as separate files.

this also lets us modify source code in a release pipeline to create new YAML to release (e.g. to publish or include in helm charts or whatever)