fix: add support for git clone as a source step
this change allow `--git-url` and an optional `--git-rev` to be passed to `build` and `resolve` commands so that instead of creating a source image bundle we can run builds from a git clone instead

I initially started this because I hit issues on my GCP cluster with source bundles and wanted to be able to try out mink #277 but on further reflection using git clone will be useful when using mink inside CI/CD pipelines triggered on a particular git revision; where creating a source image bundle is not totally necessary (though its ideal when developing locally on your laptop).

This also gives folks the option when running CI/CD pipelines with mink to use source bundle images or use git clones; there are pros and cons to both approaches and depending on your cloud infrastructure, bills, source code size and network there are trade offs for each.

The git clone step is using the git clone image from tekton + is a simplification of the tekton catalog git clone step