Refactor the bundle handling logic.
This shares the logic that handles producing a multi-arch bundle image out of kontext and surfaces an API for applying a mutation over the one (or many) `v1.Image`(s) referenced by the base image reference, and then publishing it to some tag and returning the digest.

This reduces the logic in `kontext` to essentially the file-walking logic, and enables future bundling logic to similarly assemble multi-arch base images without needing the clone this logic.

Related: https://github.com/mattmoor/mink/pull/281