Start a constants package for defining param/result names.
This starts to encapsulate some of the parameter and result naming
conventions we will make use of to duck-type tasks and pipelines
into a common package that the various utilities may draw from.

These values are not (yet) set in stone, as I'm hopeful that we can
migrate them to use the reverse-DNS style (with some support in
Tekton), to better avoid collisions.

Related: https://github.com/tektoncd/pipeline/issues/3590