fix: add a little helper for switching out the remote writes
so we can fake them out in a unit test.

Its not terribly pretty but its very handy ;)

I guess another approach is to use an interface for remote writing and let it be passed in as an option then we can use a fake inside the unit tests? I went for this approach for now as the smallest possible change